add ConfluentBlogParser
